requirements.txt
google-api-python-client
tablib
python-igraph
plotly

kubernetes
apache-airflow[gcp,kubernetes]